requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0

python-jose==3.3.0
python-dotenv==1.0.1

scikit-learn==1.3.2
joblib
pandas
numpy==1.26.4

redis==5.0.4

prometheus-fastapi-instrumentator==6.1.0

